Completed Day 03
This was an interesting exercise in simple binary operations that I
still managed to get a bit hung up on. For the generic parsing of the
input I just wrapped u16 in a local struct and implemented `FromStr`
again, as parsing the binary strings into u16 via `str.parse` isn't
really a thing as far as I could tell.

Part 1 was pretty interesting as all you needed to do was figured out
either gamma or epsilon and then you could apply a binary not op on
either to derive the other. For this part I just did gamma and iterated
each bit of each entry per iteration to increment a counts buffer for
that bit, however this was not a reusable approach for part 2.

Part 2 was a bit more involved, the filtering per bit required a bit
more clever approach to keeping count of the set bits per index. Ended
up using `FnMut` for the filter predicate for the `drain_filter` call
which is still unstable but apparently is quite a bit faster than a
typical filter due to internals or something. This part was more of
double checking that your conditions for each of the rating calculations
actually matched the expected behavior, in the end due to the hashing
mechanism it didn't matter if the calculations per rating were flipped
as long as one filtered for more commmon and the other filtered for less
common. Though my input didn't seem to run into any cases of ties where
rating dependent tie breaker logic would come into play so I am not sure
if the original solution I had would have worked for all inputs.