Day 07 done
This problem was mostly solved using guestimation and statistical
analysis of the input data set.

For day 1, it was clear that the median of the set ought to be the
correct center point, though this becomes a bit hazy if the the set is
of even count and the two prospective medians are far apart, in our case
the upper bound on the median was fine.

For day 2, this seemed immediately to be tending towards the average of
the input set, though I ran into some issues with the reality being not
quite the average, but the floor of the average rather than the rounded
value.