Day 10 complete
Pretty clearly the classic stack parser problem.

Both parts were quite easy, I just learned that indexing into the midway
point of an odd numbered list with 0 indexing is as simple as `len/2`
which I should have known but it is not a common enough issue that I
naively thought originally it should be `len/2 + 1`, so thats 30 minutes
gone but such is life.